Bugfix: passing of multiple parameters with the same name
This was completely broken. Things like passing multiple facet.field or multiple
stats.facet parameters was not working at all. Only the last one took effect
because I was using a dictionary to store the parameters. Changed this to a list
of tuples, which works equally well for the url encoder